requirements.txt
streamlit==1.28.0
python-dotenv==1.0.0
requests==2.31.0
pandas
plotly==5.17.0
beautifulsoup4==4.12.0
email-validator==2.0.0